requirements.txt
flask==2.0.2
jinja2>=3.0
tensorflow==2.6.0
keras==2.6.0
Pillow==8.3.2
numpy~=1.19.2
gunicorn==20.1.0
joblib==1.0.1
